restructure and set up grunt + webpack
