+++ package.json
@@ -11,7 +11,9 @@
     "connect-mongo": "0.8.2",
     "content-type": "^1.0.1",
     "cookie-parser": "~1.3.4",
+    "css-loader": "^0.23.1",
     "debug": "^2.2.0",
+    "elemental": "^0.5.13",
     "express": "~4.12.2",
     "express-session": "^1.13.0",
     "google-auth-library": "^0.9.7",
@@ -20,6 +22,8 @@
     "grunt-webpack": "^1.0.11",
     "jade": "~1.9.2",
     "jsx-loader": "^0.13.2",
+    "less": "^2.6.1",
+    "less-loader": "^2.2.2",
     "lwip": "0.0.8",
     "mongoose": "^4.0.1",
     "morgan": "~1.5.1",
@@ -28,8 +32,10 @@
     "nodemailer": "^2.1.0",
     "public-google-calendar": "^0.1.0",
     "react": "^0.14.7",
+    "react-addons-css-transition-group": "^0.14.7",
     "react-dom": "^0.14.7",
     "serve-favicon": "^2.2.0",
+    "style-loader": "^0.13.0",
     "webpack": "^1.12.14",
     "webpack-dev-server": "^1.14.1"
   },
