+++ public/panel/manager.html
@@ -4,7 +4,6 @@
     <meta charset="utf-8">
     <title>Manager Panel</title>
     <!-- Not present in the tutorial. Just for basic styling. -->
-    <link rel='stylesheet' href='/panel/vendor/bootstrap.min.css' />
     <link rel='stylesheet' href='/panel/css/style.css' />
     <link rel="stylesheet" href="/panel/css/base.css" />
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/2.2.0/jquery.min.js"></script>
